Move virtualbox executor integration tests to own file
